pyproject.toml
[project]
name = "powervs-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich[toml]>=0.9.1",
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.9",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "tortoise-orm[aiomysql]>=0.25.1",
    "uvicorn>=0.35.0",
]

[tool.aerich]
tortoise_orm = "core.config.tortoise_config"
location = "./migrations"
src_folder = "./."
